back_re_facial/requirements.txt
asgiref==3.8.1
contourpy==1.3.1
cycler==0.12.1
Django==5.1.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.8
fonttools==4.55.3
inflection==0.5.1
kiwisolver==1.4.8
matplotlib==3.10.0
numpy==2.2.1
opencv-python==4.10.0.84
packaging==24.2
pillow==11.1.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
uritemplate==4.1.1
